backend/requirements.txt
asgiref==3.11.0
Django==6.0
django-cors-headers==4.3.1
djangorestframework==3.16.1
pillow==12.0.0
psycopg==3.3.2
psycopg-binary==3.3.2
python-decouple==3.8
pytz==2025.2
sqlparse==0.5.4
